+++ CHANGES.md
@@ -13,9 +13,11 @@ To be released.
  -  (Libplanet.Store) Removed `ITrie.Commit()` method.  [[#3392]]
  -  (Libplanet.Store) Added `IStateStore.Commit()` method.  [[#3398]]
  -  (Libplanet.Store) Removed `IKeyValueStore.Get(IEnumerable<KeyBytes> keys)`
-    method.  [[#3362], [#3340]]
+    method.  [[#3362], [#3400]]
  -  (Libplanet.Store) Added `PathCursor` struct.  [[#3399]]
  -  (Libplanet.Store) Added `Nibbles` struct.  [[#3399]]
+ -  (Libplanet.Store) Changed `KeyBytes` to no longer implement
+    `IEquatable<byte[]>` and `IEquatable<ImmutableArray<byte>>`.  [[#3402]]
 
 ### Backward-incompatible network protocol changes
 
@@ -54,7 +56,8 @@ To be released.
 [#3392]: https://github.com/planetarium/libplanet/pull/3392
 [#3398]: https://github.com/planetarium/libplanet/pull/3398
 [#3399]: https://github.com/planetarium/libplanet/pull/3399
-[#3340]: https://github.com/planetarium/libplanet/pull/3340
+[#3400]: https://github.com/planetarium/libplanet/pull/3400
+[#3402]: https://github.com/planetarium/libplanet/pull/3402
 [RocksDB Read Only]: https://github.com/facebook/rocksdb/wiki/Read-only-and-Secondary-instances